fix DepartmentResourceTest

+++ src/test/java/resources/DepartmentsResourceTest.java
@@ -19,30 +19,35 @@ public class DepartmentsResourceTest extends ResourceTest {
 
 	@Test
 	public void testGetDepartment() {
-		Response response = target("department").path("1").register(JacksonFeature.class).request().get();
+		Response response = target("department").path("1")
+				.register(JacksonFeature.class).request().get();
 		assertEquals(Status.OK.getStatusCode(), response.getStatus());
-		DepartmentTransfer transfer = response.readEntity(DepartmentTransfer.class);
-		assertEquals(1l, transfer.getId());
+		DepartmentTransfer transfer = response
+				.readEntity(DepartmentTransfer.class);
+		assertEquals(1l, transfer.getId().longValue());
 		assertEquals("pohsun", transfer.getName());
 
 	}
 
 	@Test
 	public void testGetDepartmentWithNotFoundException() {
-		Response response = target("department").path("2").register(JacksonFeature.class).request().get();
+		Response response = target("department").path("2")
+				.register(JacksonFeature.class).request().get();
 		assertEquals(Status.NOT_FOUND.getStatusCode(), response.getStatus());
 		assertEquals("true", response.getHeaderString("safe"));
 	}
 
 	@Test
 	public void testDeleteDepartment() {
-		Response response = target("department").path("1").register(JacksonFeature.class).request().delete();
+		Response response = target("department").path("1")
+				.register(JacksonFeature.class).request().delete();
 		assertEquals(Status.OK.getStatusCode(), response.getStatus());
 	}
 
 	@Test
 	public void testDeleteDepartmentWithNotFoundException() {
-		Response response = target("department").path("2").register(JacksonFeature.class).request().delete();
+		Response response = target("department").path("2")
+				.register(JacksonFeature.class).request().delete();
 		assertEquals(Status.NOT_FOUND.getStatusCode(), response.getStatus());
 		assertEquals("true", response.getHeaderString("safe"));
 	}
@@ -51,11 +56,13 @@ public class DepartmentsResourceTest extends ResourceTest {
 	public void testUpdateDepartment() {
 		Department admin = new Department();
 		admin.setName("administrator");
-		Response response = target("department").path("1").register(JacksonFeature.class).request()
+		Response response = target("department").path("1")
+				.register(JacksonFeature.class).request()
 				.put(Entity.json(admin));
 		assertEquals(Status.OK.getStatusCode(), response.getStatus());
-		DepartmentTransfer transfer = response.readEntity(DepartmentTransfer.class);
-		assertEquals(1l, transfer.getId());
+		DepartmentTransfer transfer = response
+				.readEntity(DepartmentTransfer.class);
+		assertEquals(1l, transfer.getId().longValue());
 		assertEquals(admin.getName(), transfer.getName());
 		assertEquals("1", transfer.getManager_id());
 		assertEquals("2", transfer.getResponseto());
@@ -65,7 +72,8 @@ public class DepartmentsResourceTest extends ResourceTest {
 	public void testUpdateDepartmentWithNotFoundException() {
 		Department admin = new Department();
 		admin.setName("administrator");
-		Response response = target("department").path("2").register(JacksonFeature.class).request()
+		Response response = target("department").path("2")
+				.register(JacksonFeature.class).request()
 				.put(Entity.json(admin));
 		assertEquals(Status.NOT_FOUND.getStatusCode(), response.getStatus());
 		assertEquals("true", response.getHeaderString("safe"));
@@ -75,10 +83,12 @@ public class DepartmentsResourceTest extends ResourceTest {
 	public void testSaveDepartment() {
 		Department admin = new Department();
 		admin.setName("administrator");
-		Response response = target("department").register(JacksonFeature.class).request().post(Entity.json(admin));
+		Response response = target("department").register(JacksonFeature.class)
+				.request().post(Entity.json(admin));
 		assertEquals(Status.OK.getStatusCode(), response.getStatus());
-		DepartmentTransfer transfer = response.readEntity(DepartmentTransfer.class);
-		assertEquals(2l, transfer.getId());
+		DepartmentTransfer transfer = response
+				.readEntity(DepartmentTransfer.class);
+		assertEquals(2l, transfer.getId().longValue());
 		assertEquals(admin.getName(), transfer.getName());
 		assertEquals("1", transfer.getManager_id());
 		assertEquals("2", transfer.getResponseto());
@@ -88,19 +98,22 @@ public class DepartmentsResourceTest extends ResourceTest {
 	public void testSaveDepartmentWithDuplicateName() {
 		Department admin = new Department();
 		admin.setName("administrator");
-		Response response = target("department").register(JacksonFeature.class).request().post(Entity.json(admin));
+		Response response = target("department").register(JacksonFeature.class)
+				.request().post(Entity.json(admin));
 		assertEquals(Status.BAD_REQUEST.getStatusCode(), response.getStatus());
 	}
 
 	@Test
 	public void testFindallDepartment() {
-		Response response = target("department").register(JacksonFeature.class).request().get();
+		Response response = target("department").register(JacksonFeature.class)
+				.request().get();
 		assertEquals(Status.OK.getStatusCode(), response.getStatus());
-		Collection<DepartmentTransfer> rets = response.readEntity(new GenericType<Collection<DepartmentTransfer>>() {
-		});
+		Collection<DepartmentTransfer> rets = response
+				.readEntity(new GenericType<Collection<DepartmentTransfer>>() {
+				});
 		assertEquals(1, rets.size());
 		DepartmentTransfer transfer = rets.iterator().next();
-		assertEquals(1l, transfer.getId());
+		assertEquals(1l, transfer.getId().longValue());
 		assertEquals("pohsun", transfer.getName());
 		assertEquals("1", transfer.getManager_id());
 		assertEquals("2", transfer.getResponseto());
